backend/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
flask==3.0.3
flask-cors==4.0.1
pymupdf==1.24.0
supabase==2.3.4
pydantic==2.5.3
psycopg2-binary==2.9.9
numpy>=1.24.0

